Here is the entire solution!
I used Requests, Bs4, and smtplib in this project. They allow me to scrape the web page for price information and if the condition is true it sends an email to me notifying me of the rice change on this item. I used my Task Manager to schedule the python script to run daily that way I will never miss a sale on a product I really want.